+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - added a test utility class to send simple messages to the LED display for testing
 - mqtt variables can use the `update_template` key to define a template used to evaluate if the display should be updated. MQTT is often chatty and this can cut down on unneeded updates
 - mqtt templates can now access the previous MQTT payload via the ```previous``` variable
+- `layout.yaml` file is parsed on startup for common errors, program is halted if the schema cannot validate
 
 ## Version 1.0